fix: move actions related to font size to its own file
